+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - `Utils::digPlaces` and `Utils::decPlaces` got removed as they were no longer needed
 - removed deprecated code (`list:removeAt`, `ark` executable now replaced by `arkscript`)
 - removed `VM::getUserPointer` and `VM::setUserPointer`
+- removed `ARK_PROFILER_COUNT` define
 
 ## [3.5.0] - 2023-02-19
 ### Added